requirements.txt
pandas==2.1.4
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
crawl4ai==0.4.247
pydantic==2.10.6
